bi_platform/requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# News and data processing
newsapi-python>=0.2.6
beautifulsoup4>=4.12.0
nltk>=3.8.0
textblob>=0.17.1

# Machine Learning
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.28.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Database and storage
# sqlite3 is built into Python
sqlalchemy>=2.0.0

# API and web
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.22.0

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.0
pytz>=2023.3
